requirements.txt
# Core ML
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Reinforcement Learning
stable-baselines3>=2.0.0
gymnasium>=0.29.0

# Unsupervised Learning
scikit-learn>=1.3.0

# Game
pygame>=2.5.0

# Configuration
pyyaml>=6.0

# Optional: Monitoring
tensorboard>=2.14.0

# Optional: Enhanced clustering
# hdbscan>=0.8.33
